Step 7: Completing the Game
Added try..except block inside the play() method.
Added try...except block inside the update_board() method.
There are two places in our code that may raise an exception.

If a user enters an invalid position (like strings). So we need to wrap that code inside try..except.
If the position - 1 (position is selected by the user) by the user is not in the list. For example, our board list has 9 items. If the user selects 100 as a position, we cannot fill this position in the list and results in an exception.